feat: update table column details for interval formatting; decompose refactor SR+CR
