+++ static_report/src/components/ComparisonReport/CRTabSchemaDetails.tsx
@@ -1,0 +1,127 @@
+import {
+  Flex,
+  TableContainer,
+  Table,
+  Thead,
+  Tr,
+  Th,
+  Tbody,
+  Td,
+  Divider,
+  Text,
+} from '@chakra-ui/react';
+import { nanoid } from 'nanoid';
+
+export function CRTabSchemaDetails({ base, input }) {
+  let columns = [];
+  let mapIndex = {};
+  let i = 0;
+  let added = 0;
+  let deleted = 0;
+  let changed = 0;
+
+  Object.entries<any>(base?.columns || []).forEach(([name, column]) => {
+    mapIndex[column.name] = i;
+    columns.push({
+      name,
+      changed: true,
+      base: column,
+      input: undefined,
+    });
+    i++;
+    deleted++;
+  });
+
+  Object.entries<any>(input?.columns || []).forEach(([name, column]) => {
+    if (mapIndex.hasOwnProperty(column.name)) {
+      const index = mapIndex[column.name];
+      const isChanged = columns[index].base.schema_type !== column.schema_type;
+      columns[index] = {
+        ...columns[index],
+        input: column,
+        changed: isChanged,
+      };
+      deleted--;
+      if (isChanged) {
+        changed++;
+      }
+    } else {
+      columns.push({
+        name,
+        changed: true,
+        base: undefined,
+        input: column,
+      });
+      added++;
+    }
+  });
+
+  return (
+    <Flex direction="column">
+      <Text mb={4} p={2}>
+        Added:
+        <Text as={'span'} fontWeight={700} ml={1}>
+          {added}
+        </Text>
+        , Deleted:
+        <Text as={'span'} fontWeight={700} ml={1}>
+          {deleted}
+        </Text>
+        , Changed:{' '}
+        <Text as={'span'} fontWeight={700} ml={1}>
+          {changed}
+        </Text>
+      </Text>
+
+      <Flex justifyContent={'space-evenly'}>
+        <TableContainer width="50%">
+          <Table variant="simple">
+            <Thead>
+              <Tr>
+                <Th>Column</Th>
+                <Th>Type</Th>
+              </Tr>
+            </Thead>
+            <Tbody>
+              {columns.map((column) => (
+                <Tr
+                  key={nanoid(10)}
+                  color={column.changed ? 'red.500' : 'inherit'}
+                >
+                  <Td>{column.base?.name ?? '-'}</Td>
+                  <Td>{column.base?.schema_type ?? '-'}</Td>
+                </Tr>
+              ))}
+            </Tbody>
+          </Table>
+        </TableContainer>
+
+        <Flex justifyContent={'center'}>
+          <Divider orientation={'vertical'} />
+        </Flex>
+
+        <TableContainer width="50%">
+          <Table variant="simple">
+            <Thead>
+              <Tr>
+                <Th>Column</Th>
+                <Th>Value</Th>
+              </Tr>
+            </Thead>
+            <Tbody>
+              {columns.map((column) => (
+                <Tr
+                  key={nanoid(10)}
+                  color={column.changed ? 'red.500' : 'inherit'}
+                >
+                  <Td>{column.input?.name ?? '-'}</Td>
+                  <Td>{column.input?.schema_type ?? '-'}</Td>
+                </Tr>
+              ))}
+            </Tbody>
+          </Table>
+        </TableContainer>
+      </Flex>
+    </Flex>
+  );
+}
