+++ static_report/src/components/ComparisonReport/CRTabTestDetails.tsx
@@ -1,0 +1,91 @@
+import {
+  Flex,
+  TableContainer,
+  Table,
+  Thead,
+  Tr,
+  Th,
+  Tbody,
+  Td,
+  Text,
+} from '@chakra-ui/react';
+import { groupBy } from 'lodash';
+import { nanoid } from 'nanoid';
+import { TestStatus } from '../shared/TestStatus';
+
+export function CRTabTestDetails({ base = [], input = [], ...props }) {
+  // group by "level", "column", "name"
+  const groupedTests = groupBy(
+    [...base, ...input],
+    (test) => `${test.level}_${test.column}_${test.name}`,
+  );
+
+  const tests = Object.values(groupedTests).map((groupedTest) => {
+    let row: any = {
+      level: groupedTest[0].level,
+      column: groupedTest[0].column,
+      name: groupedTest[0].name,
+    };
+
+    (groupedTest as any).forEach((test) => {
+      if (test.from === 'base') {
+        row.base = test;
+      } else {
+        row.input = test;
+      }
+    });
+
+    return row;
+  });
+
+  if (tests.length === 0) {
+    return (
+      <Flex direction="column">
+        <Text textAlign="center">No tests available</Text>
+      </Flex>
+    );
+  }
+
+  return (
+    <TableContainer>
+      <Table variant="simple">
+        <Thead>
+          <Tr>
+            <Th>Level</Th>
+            <Th>Column</Th>
+            <Th>Assertion</Th>
+            <Th>Base Status</Th>
+            <Th>Input Status</Th>
+            <Th />
+          </Tr>
+        </Thead>
+        <Tbody>
+          {Object.values(tests).map((test) => {
+            return (
+              <Tr key={nanoid()}>
+                <Td>{test.level}</Td>
+                <Td>{test.column}</Td>
+                <Td>{test.name}</Td>
+                <Td>
+                  <TestStatus status={test.base?.status} />
+                </Td>
+                <Td>
+                  <TestStatus status={test.input?.status} />
+                </Td>
+                <Td
+                  onClick={() => {
+                    props.onDetailVisible(test);
+                  }}
+                >
+                  <Text as="span" cursor="pointer">
+                    🔍
+                  </Text>
+                </Td>
+              </Tr>
+            );
+          })}
+        </Tbody>
+      </Table>
+    </TableContainer>
+  );
+}
