+++ static_report/src/utils/index.tsx
@@ -54,9 +54,13 @@ export function getChartTooltip({ target, style = {} as any }) {
   return tooltip;
 }
 
+export type ReportAsserationStatusCounts = {
+  passed: string | number;
+  failed: string | number;
+};
 export function getReportAsserationStatusCounts(
   assertion: AssertionResult | undefined,
-) {
+): ReportAsserationStatusCounts {
   if (!assertion) {
     return { passed: '-', failed: '-' };
   }
@@ -95,22 +99,6 @@ export function getReportAsserationStatusCounts(
   };
 }
 
-export function getMissingValue(
-  column: undefined | { total?: number; non_nulls?: number },
-) {
-  if (!column) {
-    return '-';
-  }
-
-  const num = Number((column.total - column.non_nulls) / column.total) * 100;
-
-  if (Math.floor(num) === 0) {
-    return '<0.1%';
-  } else {
-    return `${num.toFixed(1)}%`;
-  }
-}
-
 export function formatReportTime(time: string) {
   const adjustForUTCOffset = (date) => {
     return new Date(
@@ -126,6 +114,13 @@ export function formatReportTime(time: string) {
   return format(adjustForUTCOffset(parseISO(time)), 'yyyy/MM/dd HH:mm:ss');
 }
 
+/**
+ *
+ * @param num number type input
+ * @param locales locale string
+ * @param options
+ * @returns a formatted string number, based on locale & options
+ */
 export function formatNumber(
   num: number,
   locales = 'en-US',
@@ -136,6 +131,32 @@ export function formatNumber(
   }
 
   return new Intl.NumberFormat(locales, options).format(num);
+}
+
+/**
+ * @param num fractional number type input
+ * @returns a formatted interval string, based on its percentage position
+ */
+export function formatIntervalMinMax(num: number) {
+  // *  should show <0.1 % if the value is between (0%, 0.1%]
+  const isLowerBound = num > 0 && num <= 0.001;
+  // *  should show >99.9% if the value is between [99.9%, 100%) .
+  const isUpperBound = num < 1 && num >= 0.999;
+
+  const formatter = (newArg = num) =>
+    formatNumber(newArg, 'en-US', {
+      style: 'percent',
+      minimumFractionDigits: 1,
+    });
+
+  if (isLowerBound) {
+    const result = formatter(0.001);
+    return `<${result}`;
+  } else if (isUpperBound) {
+    const result = formatter(0.999);
+    return `>${result}`;
+  }
+  return formatter();
 }
 
 export function extractExpectedOrActual(value) {
@@ -212,15 +233,16 @@ export function nestComparisonValueByKey(
   return result;
 }
 
+export type ComparisonAssertions = {
+  data: ComparisonReportSchema;
+  reportName: string;
+  type: 'piperider' | 'dbt';
+};
 export function getComparisonAssertions({
   data,
   reportName,
   type,
-}: {
-  data: ComparisonReportSchema;
-  reportName: string;
-  type: 'piperider' | 'dbt';
-}) {
+}: ComparisonAssertions) {
   const targets = {
     piperider: 'assertion_results',
     dbt: 'dbt_test_results',
@@ -239,6 +261,22 @@ export function getComparisonAssertions({
   return assertions;
 }
 
+//FIXME: Rename -- hard to understand
+export type ComparisonAssertionTests = {
+  passed: string | number;
+  failed: string | number;
+  tests: {
+    level: string;
+    column: string;
+    from: ComparsionSource;
+    name: string;
+    status: 'passed' | 'failed';
+    parameters: Record<string, unknown>;
+    expected: Record<string, unknown>;
+    actual: number;
+    tags: unknown[];
+  }[];
+};
 export function getComparisonAssertionTests({
   assertion,
   from,
@@ -333,23 +371,8 @@ export function transformDistributionWithLabels({ base, input, labels }) {
   return m;
 }
 
-// FIXME: Temp Typing
-export function getColumnDetails(
-  columnData?: ComparisonReportSchema['base']['tables']['ACTION']['columns'],
-) {
-  if (!columnData) {
-    return {
-      hasNoNull: false,
-      mismatch: null,
-      valid: null,
-      missing: null,
-      validOfTotal: null,
-      mismatchOfTotal: null,
-      missingOfTotal: null,
-    };
-  }
-
-  const { non_nulls, total, mismatched } = columnData as any;
+export function getColumnDetails(columnData: ColumnSchema) {
+  const { non_nulls, total, mismatched } = columnData;
 
   const hasNoNull = non_nulls === total;
 
@@ -360,6 +383,7 @@ export function getColumnDetails(
   const validOfTotal = valid / total;
   const mismatchOfTotal = mismatch / total;
   const missingOfTotal = missing / total;
+  const totalOfTotal = total / total;
 
   return {
     hasNoNull,
@@ -369,5 +393,6 @@ export function getColumnDetails(
     validOfTotal,
     mismatchOfTotal,
     missingOfTotal,
+    totalOfTotal,
   };
 }
